+++ esbuild.config.mjs
@@ -1,5 +1,5 @@
-import builtins from "builtin-modules";
 import esbuild from "esbuild";
+import { builtinModules } from "module";
 import process from "process";
 
 const banner = `/*
@@ -30,7 +30,7 @@ const context = await esbuild.context({
     "@lezer/common",
     "@lezer/highlight",
     "@lezer/lr",
-    ...builtins,
+    ...builtinModules,
   ],
   platform: "node",
   format: "cjs",